switch to float

+++ potentiometer_manager.h
@@ -3,4 +3,4 @@
 
 void potentiometer_manager_init(uint16_t potentiometer_num);
 
-void potentiometer_manager_monitor_potentiometers(uint16_t* raw_readings, void(*notify_change)(uint16_t potentiometer_id, uint8_t reading));
+void potentiometer_manager_monitor_potentiometers(float* raw_readings, void(*notify_change)(uint16_t potentiometer_id, float reading));
